+++ MATLAB/attitude_filter/helper_functions/conversion.hpp
@@ -26,12 +26,30 @@ TypedArray<T> create_from_lin_vec_arr(ArrayFactory& f, lin::Vector<T, N>* lin_ve
     return ret;    
 }
 
-template<lin::size_t R, lin::size_t C>
-TypedArray<double> create_from_lin_mat(ArrayFactory& f, lin::Matrixd<R, C> lin_mat){
+template<lin::size_t R, lin::size_t C, typename T>
+TypedArray<double> create_from_lin_mat(ArrayFactory& f, lin::Matrix<T, R, C> lin_mat){
     TypedArray<double> ret = f.createArray<double>({R, C});
     for(int r = 0; r < R; r++){
         for(int c = 0; c < C; c++)
             ret[r][c] = lin_mat(r, c);
     }
     return ret;  
+}
+
+template<lin::size_t N, typename T>
+lin::Vector<T, N> typed_array_to_lin_vec(matlab::data::TypedArray<T> typed_arr){
+    lin::Vector<T, N> ret;
+    for (int i = 0; i<N; i++) {
+        ret(i) = typed_arr[i];
+    }
+    return ret;
+}
+
+template<typename T, lin::size_t R, lin::size_t C>
+void typed_array_to_lin_mat(matlab::data::TypedArray<T> typed_arr, lin::Matrix<T, R, C>& lin_mat){
+    for( int r = 0; r<R; r++){
+        for( int c = 0; c<C; c++){
+            lin_mat(r, c) = typed_arr[r][c];
+        }
+    }
 }
