+++ MATLAB/attitude_filter/helper_functions/estimator_reset.cpp
@@ -4,40 +4,12 @@
 #include <gnc/attitude_estimator.hpp>
 #include <lin.hpp>
 
+#include "conversion.hpp"
+
 using namespace matlab::data;
 
 class MexFunction : public matlab::mex::Function {
 public:
-
-    template<lin::size_t N, typename T>
-    TypedArray<T> create_from_lin_vec(ArrayFactory& f, lin::Vector<T, N> lin_vec){
-        TypedArray<T> ret = f.createArray<T>({N, 1});
-        for(int r = 0; r < N; r++){
-            ret[r][0] = lin_vec(r);
-        }
-        return ret;
-    }
-    
-    template<lin::size_t N, typename T>
-    TypedArray<T> create_from_lin_vec_arr(ArrayFactory& f, lin::Vector<T, N>* lin_vec, size_t L){
-        TypedArray<T> ret = f.createArray<T>({N, 1, L});
-        for(int i = 0; i < L; i++){
-            for(int r = 0; r < N; r++){
-                ret[r][0][i] = (lin_vec[i])(r);
-            }
-        }
-        return ret;    
-    }
-
-    template<lin::size_t R, lin::size_t C>
-    TypedArray<double> create_from_lin_mat(ArrayFactory& f, lin::Matrixd<R, C> lin_mat){
-        TypedArray<double> ret = f.createArray<double>({R, C});
-        for(int r = 0; r < R; r++){
-            for(int c = 0; c < C; c++)
-                ret[r][c] = lin_mat(r, c);
-        }
-        return ret;  
-    }
 
     void operator()(matlab::mex::ArgumentList outputs, matlab::mex::ArgumentList inputs) {
         checkArguments(outputs, inputs);
